既存のリポジトリを置き換え
